+++ src/ScanPatron/ScanForm.test.js
@@ -4,13 +4,30 @@ import { runAxeTest } from '@folio/stripes-testing';
 import renderWithRouter from '../../test/jest/helpers/renderWithRouter';
 
 import ScanForm from './ScanForm';
+import { useReadingRoom } from '../hooks';
 
 jest.unmock('@folio/stripes/components');
 
-jest.mock('../PatronDetail', () => jest.fn(() => <div>PatronDetail</div>));
-jest.mock('../PatronAccessDetail', () => jest.fn(() => <div>PatronAccessDetail</div>));
-jest.mock('../Footer', () => jest.fn(() => <div>Footer</div>));
+jest.mock('../components/PatronDetail', () => jest.fn(() => <div>PatronDetail</div>));
+jest.mock('../components/PatronAccessDetail', () => jest.fn(() => <div>PatronAccessDetail</div>));
+jest.mock('../components/Footer', () => jest.fn(() => <div>Footer</div>));
+jest.mock('../hooks', () => ({
+  ...jest.requireActual('../hooks'),
+  useReadingRoom: jest.fn(),
+  useProfilePicConfigForTenant: jest.fn().mockReturnValue(true),
+}));
 
+const mockedReadingRoom = {
+  data: {
+    readingRooms: [
+      {
+        name: 'reading room service'
+      }
+    ]
+  },
+  refetch: jest.fn(),
+  isLoading: false,
+};
 const handleSubmit = jest.fn();
 const mockedForm = {
   change: jest.fn(),
@@ -28,22 +45,22 @@ describe('ScanForm', () => {
     handleSubmit,
     onSubmit,
     form: mockedForm,
-    scannedPatronDetails: {},
-    patronRRAPermission: {},
-    resources: {
-      userProfilePicConfig: {
-        records: [
-          { enabled: true },
-        ],
-      }
+    scannedPatronDetails: {
+      active: true
     },
+    patronRRAPermission: {},
     resetDetails,
     currUserId:'currUserId',
+    currSPId: 'currSPId',
     mutator: {},
+    loading: false,
   };
 
   describe('when scannedPatronDetails and patronRRAPermission props are set', () => {
     beforeEach(() => {
+      useReadingRoom
+        .mockClear()
+        .mockReturnValue(mockedReadingRoom);
       renderComponent(props);
     });
 
@@ -54,7 +71,7 @@ describe('ScanForm', () => {
     });
 
     it('should render component', () => {
-      expect(screen.getByText('ui-reading-room.scanPatronCard')).toBeDefined();
+      expect(screen.getByText('reading room service . ui-reading-room.scanPatronCard')).toBeDefined();
     });
 
     it('should render patron barcode field', () => {
@@ -90,6 +107,9 @@ describe('ScanForm', () => {
     };
 
     beforeEach(() => {
+      useReadingRoom
+        .mockClear()
+        .mockReturnValue(mockedReadingRoom);
       renderComponent(alteredProps);
     });
 
@@ -97,6 +117,33 @@ describe('ScanForm', () => {
       it(`should display ${text}`, () => {
         expect(screen.queryByText(`${text}`)).not.toBeInTheDocument();
       });
+    });
+  });
+
+  describe('when no reading rooms are available at the current service point', () => {
+    beforeEach(() => {
+      useReadingRoom
+        .mockClear()
+        .mockReturnValue({
+          data: {
+            readingRooms: []
+          },
+          refetch: jest.fn(),
+          isLoading: false,
+        });
+      renderComponent(props);
+    });
+
+    it('should display text to indicate no reading rooms defined at the current service point', () => {
+      expect(screen.getByText('ui-reading-room.scanPatronCard')).toBeDefined();
+    });
+
+    it('should disable input field', () => {
+      expect(screen.getByRole('textbox')).toHaveAttribute('disabled');
+    });
+
+    it('should disable enter button', () => {
+      expect(screen.getByRole('button', { name: 'ui-reading-room.enter' })).toHaveAttribute('disabled');
     });
   });
 });