+++ src/hooks/index.js
@@ -1,0 +1,2 @@
+export { useProfilePicConfigForTenant } from './useProfilePicConfigForTenant';
+export { useReadingRoom } from './useReadingRoom';