+++ src/hooks/useProfilePicConfigForTenant/index.js
@@ -1,0 +1,1 @@
+export { default as useProfilePicConfigForTenant } from './useProfilePicConfigForTenant';