+++ src/hooks/useReadingRoom/useReadingRoom.js
@@ -1,0 +1,23 @@
+import { useQuery } from 'react-query';
+
+import {
+  useNamespace,
+  useOkapiKy,
+} from '@folio/stripes/core';
+
+const useReadingRoom = (servicePointId) => {
+  const ky = useOkapiKy();
+  const [namespace] = useNamespace({ key: 'get-reading-rooms' });
+
+  const { data, refetch, isLoading } = useQuery(
+    {
+      queryKey: [namespace, servicePointId],
+      queryFn: () => ky.get(`reading-room?query=servicePoints.id==${servicePointId}`).json(),
+      enabled: !!servicePointId
+    }
+  );
+
+  return { data, refetch, isLoading };
+};
+
+export default useReadingRoom;